+++ moban/core/definitions.py
@@ -46,3 +46,16 @@ class TemplateTarget(object):
             self.output,
             self.template_type,
         )
+
+
+class Store:
+    def __init__(self):
+        self.init()
+
+    def add(self, target):
+        self.targets.append(target)
+        self.look_up_by_output[target.output] = target
+
+    def init(self):
+        self.targets = []
+        self.look_up_by_output = {}
