+++ moban/core/mobanfile/store.py
@@ -1,14 +1,3 @@
-class Store:
-    def __init__(self):
-        self.init()
-
-    def add(self, target):
-        self.targets.append(target)
-        self.look_up_by_output[target.output] = target
-
-    def init(self):
-        self.targets = []
-        self.look_up_by_output = {}
-
+from moban.core.definitions import Store
 
 STORE = Store()
